decided to not go with WPF -- too complicated for little value.  winforms is much, much easier.

+++ Burton.Lib.Graph/GraphEdge.cs
@@ -28,9 +28,9 @@ namespace Burton.Lib.Graph
         /// <summary>
         /// The cost of traversing the Edge
         /// </summary>
-        public float EdgeCost;
+        public double EdgeCost;
 
-        public GraphEdge(int FromNodeIndex, int ToNodeIndex, float EdgeCost)
+        public GraphEdge(int FromNodeIndex, int ToNodeIndex, double EdgeCost)
         {
             this.FromNodeIndex = FromNodeIndex;
             this.ToNodeIndex = ToNodeIndex;
